rest_api/requirements.txt
django>=4.2
psycopg2-binary
djangorestframework
django-filter
python-dotenv
numpy
